Bill S7269 version original - Kevin S. Parker
BILL NUMBER: S7269 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing the New York state council on mental health emergency and crisis response PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bills is to study the effects of a non-law enforce- ment emergency mental health response unit. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The mental hygiene law is amended by adding a new section 7.