Bill S7265 version original - Kevin S. Parker
BILL NUMBER: S7265 SPONSOR: PARKER TITLE OF BILL: An act to amend the public officers law, in relation to residency requirements for paid firefighters in a city with a population of one million or more PURPOSE OR GENERAL IDEA OF BILL: To establish a residency requirement for the FDNY. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends the public officers law to provide an exemption against residency requirement for cities with a population of one million or more. Section 2.