Bill S7279 version original - Kevin S. Parker
BILL NUMBER: S7279 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to establishing a duty to report an unauthorized use of force by a police officer PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to require police officers to intervene and report incidents of police misconduct. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends the executive law to require police officers to inter- vene on and report instances of unauthorized use of force.