Bill S7347 version original - Kevin S. Parker
BILL NUMBER: S7347 SPONSOR: PARKER TITLE OF BILL: An act requiring the empire state development corporation to conduct a study on the economic impact of the Labor Day Parade (West Indian Carni- val) PURPOSE: To conduct a study on the economic impact of the West Indian Carnival. SUMMARY OF PROVISION: The proposed legislation would require that Empire State Development to conduct a study of the economic impact of the West Indian Carnival.