Bill S7348 version original - Kevin S. Parker
BILL NUMBER: S7348 SPONSOR: PARKER TITLE OF BILL: An act to amend the election law and the correction law, in relation to authorizing voting by incarcerated individuals in correctional facili- ties; and to repeal certain provisions of the election law relating thereto PURPOSE: This legislation would extend certain inmates in correctional facilities the ability to vote in elections. SUMMARY OF PROVISIONS: Section 1 - Amends the election law as required.