Bill S7320 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S7320 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the penal law, in relation to increasing the penalties for certain crimes involving tampering with a witness or intimidating a victim or witness PURPOSE: To help protect criminal witnesses by elevating the penalties for crimes of tampering with or intimidating a witness. SUMMARY OF PROVISIONS: Section 1: Amends section 215.