Bill S7354 version original - Michelle Hinchey
BILL NUMBER: S7354 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the cannabis law, in relation to extending certain authorizations to minimally process and distribute cannabis products PURPOSE: The purpose of this bill is to extend the authorization of Chapter 18 of 2022 for one additional year. SUMMARY OF PROVISIONS: Section 1 would amend section 88-c to the Cannabis Law to extend a conditional adult-use cultivator license for one year.