Bill S7349 version original - Kevin S. Parker
BILL NUMBER: S7349 SPONSOR: PARKER TITLE OF BILL: An act directing the New York state comptroller to conduct a feasibility study on implementing a universal basic income program PURPOSE: This bill will authorize and direct the Office of the New York State Comptroller to study and report on the feasibility of implementing a universal basic income program in New York State. SUMMARY OF PROVISIONS: 1.