Bill S7322 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S7322 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the criminal procedure law, in relation to authorizing crimes committed by members of street gangs or a criminal enterprise to be eligible for bail PURPOSE:: To allow judges discretion in setting bail if there is a reasonable cause to believe such crime is connected to criminal street gang or criminal enterprise activity. SUMMARY OF PROVISIONS:: Section 1: Amends Section 510.