Bill S7272 version original - Kevin S. Parker
BILL NUMBER: S7272 SPONSOR: PARKER TITLE OF BILL: An act to amend the penal law, in relation to the justification for the use of force in a homicide by a peace officer or police officer PURPOSE OF BILL: The purpose of the bill is to heighten the justification for the use of force by peace officer or police officer. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amended the penal law by adding two new sections 35.35 and 35.40 Section 2.