Bill S7323 version original - Patricia Canzoneri-Fitzpatrick
BILL NUMBER: S7323 SPONSOR: CANZONERI-FITZPATRICK TITLE OF BILL: An act to amend the public health law, in relation to Medicaid reimbursement rates for early intervention services PURPOSE:: Relates to early intervention reimbursement rates. SUMMARY OF PROVISIONS:: Section 1. Subdivision 2 of section 2557 of the public health law, as amended by chapter 820 of the laws of 2021, is amended to read as follows: 2.