Bill S7294 version original - Michelle Hinchey
BILL NUMBER: S7294 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the public service law, in relation to certain penalties PURPOSE OR GENERAL IDEA OF BILL: The penalties included in Section 25 and 25-a of the public service law have not been updated since they were introduced in 1986. This bill increases the penalties by 272% to account for historic inflation, and indexes them on an annual basis going forward.