Bill S3444A version A - Pamela Helming
BILL NUMBER: S3444A SPONSOR: HELMING TITLE OF BILL: An act to amend the tax law, in relation to authorizing Wayne county to impose a hotel, motel or seasonal rental occupancy tax; and providing for the repeal of such provisions upon expiration thereof PURPOSE: Authorizes the county of Wayne to impose a hotel or motel room or seasonal rental occupancy tax. SUMMARY OF PROVISIONS: Section 1: Adds a new section 1202-hh to the new tax law.