Bill S7396 version original - Anthony H. Palumbo
BILL NUMBER: S7396 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the criminal procedure law, in relation to granting constables on Fishers Island in the town of Southold, Suffolk county, peace officer status PURPOSE: To grant peace officer status to the constables of Fisher's Island in the town of Southold. SUMMARY OF PROVISIONS: Section 1. Section 2.10 of the criminal procedure law is amended by adding a new subdivision 87. Section 2. Effective date.