Bill S7392A version A - Zellnor Myrie
BILL NUMBER: S7392A SPONSOR: MYRIE TITLE OF BILL: An act to amend the general business law, in relation to the dangers to safety and health and the creation of a public nuisance caused by the sale, manufacturing, distribution, importing, and marketing of firearms PURPOSE:: To make technical fixes. SUMMARY OF PROVISIONS:: Section 1 amends subdivision 6 of section 898-a of the general business law by adding chapter 237 of the laws of 2021 which defines the term, "qualified product".