Bill S7382 version original - Brad Hoylman-Sigal
BILL NUMBER: S7382 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the executive law, in relation to adding gender identity or expression as a protected class of the human rights law PURPOSE OF THE BILL: Adding Gender Identity or Expression as a Protected Class to the Intern Provision of the Human Rights Law. SUMMARY OF PROVISIONS: Section 1 of the bill makes a technical correction to the Human Rights Law to renumber paragraphs.