Bill S5650A version A - John Liu
BILL NUMBER: S5650A SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to expanding access to advanced courses to improve educational equity PURPOSE OR GENERAL IDEA OF BILL: This bill seeks to expand access to advanced courses to improve educa- tional equity. SUMMARY OF PROVISIONS: Section 1 adds a new section 817 to education law that provides for notice to parents on the availability of and information on advanced courses.