Bill S7365B version original - Brad Hoylman-Sigal
BILL NUMBER: S7365 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law and the general business law, in relation to rapid fire modification devices PURPOSE: To add common conversion devices to the definition of rapid-fire modifi- cation devices for the purposes of firearm offenses and to require firearm sellers to take reasonable precautions to prevent the installa- tion or use of such devices. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends section 265.