Bill S7388A version original - Kevin S. Parker
BILL NUMBER: S7388 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to permissible activities for athletic associations, conferences, or other groups or organizations with authority over intercollegiate athletics and the definition of student-athlete PURPOSE: To enhance New York State's ability to attract student-athletes, and to provide student-athletes with better name, images, and likeness opportu- nities. SUMMARY: Section 1.