Bill S942A version A - Jamaal Bailey
BILL NUMBER: S942A SPONSOR: BAILEY TITLE OF BILL: An act to amend the education law, in relation to requiring licensure of athletic trainers; and to amend the social services law, in relation to making athletic trainers mandatory reporters of suspected child abuse or maltreatment PURPOSE: To create licensure for the profession of athletic training. SUMMARY: Section 1. Amends section 8351 of the education law. Section 2. Amends section 8352 of the education law. Section 3.