Bill S7364A version original - Brad Hoylman-Sigal
BILL NUMBER: S7364 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the penal law and the criminal procedure law, in relation to the regulation of three-dimensional printed firearms PURPOSE: To prohibit the use of three-dimensional printers to manufacture ghost guns, silencers, magazines, or other firearm parts and to prohibit the sale, distribution and distribution of the digital instructions to do the same. SUMMARY OF SPECIFIC PROVISIONS: Section 1 amends section 265.