Bill S4830C version B - Timothy M. Kennedy
BILL NUMBER: S4830B SPONSOR: KENNEDY TITLE OF BILL: An act to amend the public authorities law, in relation to conducting a highway and depot charging needs evaluation PURPOSE: This legislation amends the public authorities and public service laws to require the development of a highway and depot charging action plan to meet the state's clean transportation goals. SUMMARY OF PROVISIONS: Section 1. Sets for the legislative findings. § 2.