Bill S7392A version original - Zellnor Myrie
BILL NUMBER: S7392 SPONSOR: MYRIE TITLE OF BILL: An act to amend the general business law, in relation to ensuring all firearm industry members are responsible for such member's unreasonable conduct PURPOSE: To make technical fixes. SUMMARY OF PROVISIONS: Section 1 amends subdivision 6 of section 898-a of the general business law by adding chapter 237 of the laws of 2021 which defines the term, "qualified product". Section 2 provides for the effective date.