Bill S2935D version C - Kevin S. Parker
BILL NUMBER: S2935C SPONSOR: PARKER TITLE OF BILL: An act enacting the "just energy transition act" PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to provide a blueprint to guide the replace- ment and redevelopment of New York State's fossil fuel facilities and their sites by 2030. SUMMARY OF PROVISIONS: Section 1 identifies the title of the bill. Section 2 is the legislative findings.