Bill S7401 version original - Andrew J. Lanza
BILL NUMBER: S7401 SPONSOR: LANZA TITLE OF BILL: An act to amend the education law, in relation to prohibiting the hous- ing of migrants in schoolhouses or on such grounds PURPOSE OF BILL:: Prohibits the housing of migrants in schoolhouses or on such grounds. SUMMARY OF SPECIFIC PROVISIONS:: Section 1: Section 2590-h of the education law is amended by adding a new subdivision 57. Section 2: Is the effective date.