Bill S6027 version A - William Weber
BILL NUMBER: S6027A SPONSOR: WEBER TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Dr. Robert C. "Doc" Schelin Memorial Highway" PURPOSE:: Designates all that portion of the state highway system constituting State Route 340 between Thorpe Drive and State Route 303 in the town of Orangetown, County of Rockland, shall be designated and be known as the "Dr. Robert C. "Doc" Schelin Memorial Highway.