Bill S5261A version A - Julia Salazar
BILL NUMBER: S5261A SPONSOR: SALAZAR TITLE OF BILL: An act to amend the education law, in relation to prohibiting corporal punishment in schools PURPOSE OR GENERAL IDEA OF BILL: This bill clarifies that corporal punishment in an educational setting constitutes child abuse. SUMMARY OF SPECIFIC PROVISIONS: Section one amends subdivision 1 of section 1125 of the education law to add corporal punishment to the definition of "child abuse".