Bill S7421 version original - Robert Ortt
BILL NUMBER: S7421 SPONSOR: ORTT TITLE OF BILL: An act to amend the veterans' services law, in relation to the veterans employment task force PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is place into the Veterans' Services Law, the task force to study and improve veteran employment opportunities. SUMMARY OF PROVISIONS: Section 1: Adds a new Section 3-a of the Veterans' Services Law. Section 2: Effective date.