Bill S7429 version original - Monica Martinez
BILL NUMBER: S7429 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the executive law, in relation to the definition of small business PURPOSE: The purpose of this bill is to align New York State's minority and women-owned business (MWBE) employee size standards with the federal standards for small businesses. SUMMARY OF PROVISIONS: Section 1.