Bill S893A version A - Brad Hoylman-Sigal
BILL NUMBER: S893A SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the court of claims act, in relation to proof of claims for unjust conviction and imprisonment PURPOSE: To allow for claims for unjust conviction based upon actual innocence. SUMMARY OF SPECIFIC PROVISIONS: Sections 1 and 2 amend Court of Claims Act § 8-b to make it gender neutral and to authorize a claimant to bring a wrongful conviction claim when the conviction was reversed or vacated due to actual innocence.