Bill S5521A version A - Jeremy Cooney
BILL NUMBER: S5521A SPONSOR: COONEY TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to distinctive license plates issued to congressional medal of honor recipients PURPOSE:: This bill would align the eligibility requirements for the issuance of Congressional Medal of Honor license plates, with existing military honors such as the Purple Heart. Presently, the spouses of Congressional Medal of Honor recipients are ineligible to apply for this distinctive license plate.