Bill S1780A version A - Jamaal Bailey
BILL NUMBER: S1780A SPONSOR: BAILEY TITLE OF BILL: An act to incentivize the purchase of local produce in supermarkets and grocery stores through the FreshConnect program PURPOSE OR GENERAL IDEA OF BILL: To expand the reach of the FreshConnect program across the state and increase access to affordable fruits and vegetables to. New Yorkers in need.