Bill S7432A version original - Rachel May
BILL NUMBER: S7432 SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to the practice of physi- cians licensed in another state or territory of the United States pursu- ant to a reciprocal agreement PURPOSE: To authorize physicians licensed in another state or territory of the United States to practice in New York State pursuant to a reciprocal agreement. SUMMARY OF SPECIFIC PROVISIONS: Section 1, of the bill amends the Education.