Bill S7431A version original - Robert Rolison
BILL NUMBER: S7431 SPONSOR: ROLISON TITLE OF BILL: An act to amend chapter 371 of the laws of 2020 amending the tax law relating to authorizing the town of Newburgh to impose a hotel and motel tax, in relation to the effectiveness thereof PURPOSE: This legislation extends the town of Newburgh's hotel and motel tax through December 23. 2026.