Bill S7434A version original - Robert Rolison
BILL NUMBER: S7434 SPONSOR: ROLISON TITLE OF BILL: An act to amend the public officers law and the village law, in relation to residency requirements for the village clerk/treasurer and the deputy clerk for the village of Cold Spring PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to waive the residency requirements for the village clerk/treasurer and the deputy clerk for the village of Cold Spring. SUMMARY OF PROVISIONS: Section 1.