Bill S7433 version original - Monica Martinez
BILL NUMBER: S7433 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the transportation law, in relation to maximum grant awards for airport improvement and revitalization PURPOSE: The purpose of this bill is to authorize the commissioner of the depart- ment of transportation to increase the maximum aviation capital grant award from $2.5 million to up to $5 million. SUMMARY OF PROVISIONS: Section 1.