Bill S916B version A - Jeremy Cooney
BILL NUMBER: S916A SPONSOR: COONEY TITLE OF BILL: An act to amend the banking law, in relation to providing for income access services in the state PURPOSE OR GENERAL IDEA OF BILL: Provides for income access services in the state. SUMMARY OF SPECIFIC PROVISIONS: Section 374-aa defines terms used in the article.