Bill S7459 version original - Monica Martinez
BILL NUMBER: S7459 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the education law, in relation to information to be provided by a veterinarian initially prescribing or dispensing a drug to a dog, cat, or rabbit; and to amend Buoy's law, in relation to the effectiveness thereof PURPOSE: The purpose of this bill is to provide technical amendments to chapter 678 of the Laws of 2022. SUMMARY OF PROVISIONS: Section 1.