Bill S7454 version original - Monica Martinez
BILL NUMBER: S7454 SPONSOR: MARTINEZ TITLE OF BILL: An act to repeal title 16 of article 8 of the public authorities law, relating to the Suffolk county judicial facilities agency PURPOSE: The purpose of this bill is to repeal article 8, title 16 of the public authorities law which will dissolve the Suffolk county Judicial Facili- ties Agency. SUMMARY OF PROVISIONS: Section 1. repeals title 16 article 8 of the public authorities law. Section 2. is the enacting clause.