Bill S7453 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S7453 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the penal law, in relation to including video lottery gaming in certain gambling offenses PURPOSE OF THE BILL: This bill extends coverage of the Penal Law for certain casino-related crimes to similar conduct at video lottery gaming facilities. SUMMARY OF PROVISIONS: Section 1 of the bill amends subdivisions 13, 22 and 27 of section 225.