Bill S809A version A - José M. Serrano
BILL NUMBER: S809A SPONSOR: SERRANO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to prohibiting littering and dumping on state park lands PURPOSE OR GENERAL IDEA OF BILL: To create a deterrent to littering and dumping in State Parks and Historic Sites. SUMMARY OF SPECIFIC PROVISIONS: Section one will add a new Section 13.12 to the parks, recreation and historic preservation law to prohibit littering on park lands.