Bill S7473 version original - Jamaal Bailey
BILL NUMBER: S7473 SPONSOR: BAILEY TITLE OF BILL: An act to amend the insurance law, in relation to discrimination PURPOSE OR GENERAL IDEA: The purpose of the bill is to prevent insurers from discriminating against policy holders with respect to various socioeconomic factors, as well as to prevent insurance redlining. SUMMARY OF SPECIFIC PROVISIONS: Section one of the bill amends the relevant provisions of the insurance law to expand the list of protected groups under the existing statute.