Bill S7220A version A - Kevin S. Parker
BILL NUMBER: S7220A SPONSOR: PARKER TITLE OF BILL: An act to amend the real property actions and proceedings law, in relation to registration of mortgages in default PURPOSE: To provide municipalities with a fee structure for the maintenance of foreclosure registration databases. SUMMARY OF PROVISIONS: Section 1 amends the law to allow for the imposition of a fee of up to five hundred dollars on an annual basis. Section two provides for an immediate effective date.