Bill S7431A version A - Robert Rolison
BILL NUMBER: S7431A SPONSOR: ROLISON TITLE OF BILL: An act to amend chapter 371 of the laws of 2020 amending the tax law relating to authorizing the town of Newburgh to impose a hotel and motel tax, in relation to the effectiveness thereof PURPOSE: This legislation extends the town of Newburgh's hotel and motel tax through December 23. 2025. SUMMARY OF PROVISIONS: Section 1: amends chapter 371 of the tax law to extend the hotel and motel tax in the Town of Newburgh to December 23, 2025.