Bill S760B version B - John Liu
BILL NUMBER: S760B SPONSOR: LIU TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to prohibiting reckless driving in parking lots PURPOSE: The purpose of this bill is to address a dangerous trend of motorists engaging in reckless behavior in parking lots that are open to the public. SUMMARY OF PROVISIONS: Section 1 amends the definition of reckless driving to include driving or using a motor vehicle in parking lots. Section 2 provides the effective date.