Bill S7470 version original - Anthony H. Palumbo
BILL NUMBER: S7470 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the domestic relations law, in relation to the rights of grandparents with respect to visitation rights with minor children PURPOSE: The purpose of this bill is to require that grandparents be granted legal standing by a judge being considered for visitation rights of a minor child. SUMMARY OF PROVISIONS: Section 1.