Bill S6350B version B - Rachel May
BILL NUMBER: S6350B SPONSOR: MAY TITLE OF BILL: An act to amend the education law, in relation to enacting the "freedom to read act" SUMMARY OF PROVISIONS: Amends sections 276 and 283 of the Education Law to ensure that school libraries and librarians are empowered to curate the widest array of developmentally appropriate collections for students to access. The B-print of this legislation changes the section of law amended. JUSTIFICATION: "Until I feared I would lose it, I never loved to read.