Bill S7488 version original - Sean Ryan
BILL NUMBER: S7488 SPONSOR: RYAN TITLE OF BILL: An act to amend the judiciary law and the state finance law, in relation to the attorney licensing fund This measure is being introduced at the request of the Judiciary. This measure would amend section 465(1) of the Judiciary Law to codify in the Consolidated Laws a provision directing the deposit of bar exam- ination fees into the Attorney Licensing Fund.