Bill S6348B version B - Rachel May
BILL NUMBER: S6348B SPONSOR: MAY TITLE OF BILL: An act to amend the real property law, in relation to enacting the "tenants organizing act" SUMMARY OF PROVISIONS: Expands section 230 of the real property law to include definitions and protected activities of tenants' organizations, a process for establish- ing and being recognized, and a requirement for good faith engagement between property owners and tenant organizations. The B-print makes technical changes.