Bill S7482 version original - Sean Ryan
BILL NUMBER: S7482 SPONSOR: RYAN TITLE OF BILL: An act to amend the civil practice law and rules, in relation to waiver of privilege PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to bring New York State's CPLR rule regard- ing waiver of attorney client privilege into conformity with the federal rule of evidence regarding waiver of attorney client privilege. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Creates a new section of CPLR, section 4550.