Bill S7119A version A - Jessica Ramos
BILL NUMBER: S7119A SPONSOR: RAMOS TITLE OF BILL: An act to amend the general municipal law, in relation to the purchase of apparatus, materials, equipment or supplies PURPOSE OR GENERAL IDEA OF BILL: The purpose of the bill is to prohibit the use of piggyback contracting for construction services. SUMMARY OF PROVISIONS: Amend section 103(16) of the general municipal law to scale back the scope of services which may be utilized through such procurement method.