Bill S6476B version A - Kristen Gonzalez
BILL NUMBER: S6476A SPONSOR: GONZALEZ TITLE OF BILL: An act directing a state agency telework report PURPOSE: The purpose of the report is to determine how state agencies are allow- ing their workers to telecornmute and the effects of telecommuting by state workers. The study would require state agencies to submit informa- tion on their official telecommuting policy as well as data on how many workers and which job titles are eligible for telework.