Bill S5983A version A - Monica Martinez
BILL NUMBER: S5983AREVISED 6/2/23 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public health law, in relation to requiring each certificate of death to also be filed with each board of elections with- in the state PURPOSE: The purpose of this bill is to require each board of elections receives the monthly reports of death certificates from the departments of health. SUMMARY OF PROVISIONS: Section 1.