Bill S7504 version original - Jamaal Bailey
BILL NUMBER: S7504 SPONSOR: BAILEY TITLE OF BILL: An act to amend the cannabis law, in relation to expanding who can hold membership in an adult-use cooperative license PURPOSE OR GENERAL IDEA: The purpose of this bill is to expand who can hold membership in an adult-use cooperative license. SUMMARY OF SPECIFIC PROVISIONS: Section one of the bill amends section 70 of the cannabis law to strike out the language pertaining to "natural person." Section two of the bill sets forth the effective date.