Bill S7505 version original - Jamaal Bailey
BILL NUMBER: S7505 SPONSOR: BAILEY TITLE OF BILL: An act to amend the criminal procedure law, in relation to making tech- nical corrections regarding the unlawful possession and sale of cannabis PURPOSE: OR GENERAL IDEA: To correct drafting errors to the MRTA. SUMMARY OF SPECIFIC PROVISIONS: Section one of the bill makes a technical change to section 440.46-a of the criminal procedure law. Section two of the bill sets forth the effective date.