Bill S7437A version A - Dean Murray
BILL NUMBER: S7437A SPONSOR: MURRAY TITLE OF BILL: An act to amend the real property law, in relation to lease-hold retire- ment communities PURPOSE: This legislation defines a lease-hold retirement community and proscribes rules and regulations for ownership and operation under the Real Property Law. It also provides that homes in a leasehold retirement community would be subject to provisions of Section 233-b of the Real Property Law relating to rent increases.