Bill S7501 version original - Monica Martinez
BILL NUMBER: S7501 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the veterans' services law, in relation to veterans health care PURPOSE OR GENERAL IDEA OF BILL: This bill calls for the New York State Department of Veterans' Services, in consultation with the department of health, to undertake a study regarding veteran healthcare in New York State. SUMMARY OF PROVISIONS: Section 1: amends the veterans' services law by adding a new section 18-a.