Bill S6519A version A - Brian Kavanagh
BILL NUMBER: S6519A SPONSOR: KAVANAGH TITLE OF BILL: An act to amend the election law, in relation to establishing a deadline for changing the location of a polling place for an early voting period PURPOSE OF BILL: Early voting place change notification. SUMMARY OF PROVISIONS: To amend the election law, in relation to establishing a deadline for changing the location of a polling palace for an early voting period.