Bill S7395C version B - William Weber
BILL NUMBER: S7395B SPONSOR: WEBER TITLE OF BILL: An act in relation to authorizing the assessor of the town of Ramapo, county of Rockland, to accept an application for a real property tax exemption from Sister Servants of Mary Immaculate Inc. PURPOSE: This act authorizes the assessor of the Town of Ramapo to accept an application for a real property tax exemption from Sister Servants of Mary Immaculate Inc.