Bill S1604F version C - Jessica Ramos
BILL NUMBER: S1604C SPONSOR: RAMOS TITLE OF BILL: An act to amend the labor law, in relation to regulating the temperature of all indoor and outdoor worksites PURPOSE: This legislation regulates the high temperature in indoor and outdoor worksites within the scope of the article. SUMMARY OF PROVISIONS: Section 1. This act shall be known and may be cited as the "Temperature Extreme Mitigation (TEMP) act" Section 2. Legislative Findings and Intent Section 3.