Bill S7513 version original - Monica Martinez
BILL NUMBER: S7513 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the state finance law, in relation to making technical corrections to a definition for purposes of the excelsior linked deposit program PURPOSE:: The purpose of this bill is to provide technical corrections to the definition of certified Service-Disabled Veteran Owned Business Enter- prises within the Excelsior Linked Deposit program. SUMMARY OF PROVISIONS:: Section 1.