Bill S7527 version original - Iwen Chu
BILL NUMBER: S7527 SPONSOR: CHU TITLE OF BILL: An act to amend the education law, in relation to restricted clinical laboratory licenses PURPOSE: To amend the molecular diagnosis restricted license related to consol- idating the license to streamline all aspects of molecular testing. SUMMARY OF SPECIFIC PROVISIONS: Section 1 of the bill amends section 8610 of the education law by amend- ing the molecular diagnosis restricted license. Section 2 of the bill is the effective date.