Bill S7518 version original - Neil D. Breslin
BILL NUMBER: S7518 SPONSOR: BRESLIN TITLE OF BILL: An act to amend the judiciary law, in relation to authorizing the state board of law examiners to procure quarters for such board in the county of Albany This measure is being introduced at the request of the Acting Chief Administrative Judge on behalf of the State Board of Law Examiners ("State Board"). This measure would amend section 461 of the Judiciary Law to permit location of the State Board's office anywhere in Albany County.