Bill S7525 version original - Brad Hoylman-Sigal
BILL NUMBER: S7525 SPONSOR: HOYLMAN-SIGAL TITLE OF BILL: An act to amend the judiciary law, in relation to the compensation of persons who serve as jurors This measure is being introduced at the request of the Judiciary. This measure would amend sections 519 and 521(a) of the Judiciary Law to increase the per diem paid persons who serve as trial and grand jurors in the courts of the Unified Court System.