Bill S7533A version original - Michelle Hinchey
BILL NUMBER: S7533 SPONSOR: HINCHEY TITLE OF BILL: An act to amend the agriculture and markets law, in relation to the Nourish New York program PURPOSE OR GENERAL IDEA OF BILL: The purpose of the legislation is to direct monies appropriated for the Nourish New York program to be distributed as was intended when the law was established. The bill clarifies that the funds must be distributed to the ten regional food banks, which may suballocate funding to food relief organizations.