Bill S7548 version original - Zellnor Myrie
BILL NUMBER: S7548 SPONSOR: MYRIE TITLE OF BILL: An act to amend the criminal procedure law, in relation to motions to vacate judgment; and to repeal certain provisions of such law relating thereto SUMMARY OF SPECIFIC PROVISIONS: Section 1 adds sections 440.00 to the Criminal Procedure Law, which defines the term "applicant" as a person previously convicted of a crime who is applying for relief under this article. Section 1 also adds section 440.