Bill S513B version B - Kevin Thomas
BILL NUMBER: S513B SPONSOR: THOMAS TITLE OF BILL: An act to amend the education law, in relation to establishing certain school holidays PURPOSE: The purpose of the bill is to avail students from different religious and cultural backgrounds to celebrate their traditional festivities with their family and community. SUMMARY OF PROVISIONS: Section 1 of the bill amends section 3210 of the education law section and adds a new section 3204-a.