Bill S7552 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S7552 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law, in relation to establishing Diwali as a school holiday in certain public schools PURPOSE OR GENERAL IDEA OF BILL:: To designate Diwali as a school holiday in the city of New York. SUMMARY OF PROVISIONS:: Section 1 amends the education law by adding a new section 2586-a Section 2 amends Subdivision 2-a of section 3635 of the education law. Section 3 sets the effective date.