Bill S7545 version original - James Tedisco
BILL NUMBER: S7545 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to adjudi- cations and owner liability for a violation of traffic-control signal indications in the city of Schenectady; to amend the public officers law, in relation to accessing records; and providing for the repeal of certain provisions upon expiration thereof PURPOSE: To allow the city of Schenectady to establish a red-light camera ticket- ing program. SUMMARY OF PROVISIONS: Section 1.