Bill S7543B version original - Kristen Gonzalez
BILL NUMBER: S7543 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the state technology law, in relation to automated deci- sion-making by state agencies PURPOSE: This bill would prohibit the unauthorized use of automated decision- making technologies by state agencies or entities on behalf of an agency and, when authorized, require the publiation of an impact assesment. SUMMARY OF PROVISIONS: Section 1.