Bill S7538 version original - James Skoufis
BILL NUMBER: S7538 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the village law, in relation to the incorporation of villages PURPOSE: Changes the requirements for village incorporation to ensure that village incorporations are properly assessed and the petition process fully considers the fiscal, service, and taxation ramifications of the potential incorporation of a new village. SUMMARY OF PROVISIONS: Section 1 changes the minimum size for village incorporation from 500 to 2,000 residents.