Bill S7216C version A - Kristen Gonzalez
BILL NUMBER: S7216A SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the tax law, in relation to establishing a tax on noise emissions from non-essential helicopter and seaplane flights in cities with a population of one million or more PURPOSE: This bill will impose a noise tax on non-essential helicopter and seap- lane flights in New York City. The revenue from these taxes will be dedicated to and administered by the State's Environmental Protection Fund. SUMMARY OF PROVISIONS: Section 1.