Bill S7560 version original - Anthony H. Palumbo
BILL NUMBER: S7560 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the real property tax law, in relation to extending certain provisions related to the affordable New York housing program PURPOSE: To extend the 421-a tax exemption program's start date and completion date. SUMMARY OF PROVISIONS: Sections 1-4 of the 421-a section of the real property tax law are amended to extend the project commencement date to December 31, 2030 and the project completion date to December 31, 2034. Section 5.