Bill S4981B version B - Samra Brouk
BILL NUMBER: S4981B SPONSOR: BROUK TITLE OF BILL: An act to amend the public health law, in relation to requiring hospi- tals to adopt, implement and periodically update standard protocols for the management of fetal demise PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to ensure expected mothers who encounter fetal deaths get the treatment they need, especially in the event of an emergency. SUMMARY OF SPECIFIC PROVISIONS: Section 1.