Bill S7558 version original - Robert Jackson
BILL NUMBER: S7558 SPONSOR: JACKSON TITLE OF BILL: An act to amend the education law, in relation to requiring the depart- ment of education to develop school health and mental health profes- sionals to student ratios in public schools PURPOSE: This legislation would require the department of education to develop school health and mental health professionals-to-student ratios in public schools. SUMMARY OF PROVISIONS: Section 1 sets forth the legislative intent.