Bill S7554 version original - Dan Stec
BILL NUMBER: S7554 SPONSOR: STEC TITLE OF BILL: An act to amend the family court act, in relation to the number of judg- es of the family court in St. Lawrence county PURPOSE: To create an additional post of family court judge in the county of St. Lawrence SUMMARY OF PROVISIONS: Section 1 - Amends section 131 of the family court act to add one addi- tional office of family court judge in the county of St. Lawrence. Section 2 - Provides effective date JUSTIFICATION: On June 5, 2023, the St.