Bill S7362A version A - Jeremy Cooney
BILL NUMBER: S7362A SPONSOR: COONEY TITLE OF BILL: An act to amend the general municipal law, in relation to electronic bidding PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to authorize the City of New York, and agen- cies within, to require electronic bidding. SUMMARY OF PROVISIONS: Sections one and two of this bill would provide that the appropriate officer, board or agency of the City of New York may require bids and offers may be submitted electronically.