Bill S6349A version A - Neil D. Breslin
BILL NUMBER: S6349A SPONSOR: BRESLIN TITLE OF BILL: An act to amend the executive law, in relation to adding North Chuctan- unda creek to the definition of inland waterways PURPOSE OR GENERAL IDEA OF BILL: Adding North Chuctanunda Creek to the list of New York State's major creeks. SUMMARY OF PROVISIONS: Subdivision 4 of section 911 of the executive law is amended to add North Chuctanunda Creek to the list of New York State's major creeks.