Bill S7574 version original - Joseph P. Addabbo Jr.
BILL NUMBER: S7574 SPONSOR: ADDABBO TITLE OF BILL: An act to amend the education law, in relation to establishing Diwali as a school holiday in certain public schools PURPOSE OR GENERAL IDEA OF BILL: To declare Diwali as a school holiday in the city of New York. SUMMARY OF PROVISIONS: Section 1 adds a new section 2586-a of education law to declare Diwali as a public school holiday for all public schools in New York City and require all the public schools to not be in session on such day.