Bill S1051A version A - Jeremy Cooney
BILL NUMBER: S1051A SPONSOR: COONEY TITLE OF BILL: An act to amend the executive law, in relation to establishing the New York Asian American and Pacific Islander commission SUMMARY OF PROVISIONS: Section 1 adds a new section 295-a to the executive law which estab- lishes the Asian American and Pacific Islander commission. Section 2 is the effective date.