Bill S7581 version original - Kevin S. Parker
BILL NUMBER: S7581 SPONSOR: PARKER TITLE OF BILL: An act to amend the workers' compensation law, in relation to providing disability benefits for pregnancy loss PURPOSE OR GENERAL IDEA OF BILL: To provide temporary disability benefits for pregnancy loss. SUMMARY OF PROVISIONS: The bill would amend the Workers' Compensation Law to include pregnancy loss as a qualifying condition for the receipt of temporary disability benefits.