Bill S7583 version original - John Liu
BILL NUMBER: S7583 SPONSOR: LIU TITLE OF BILL: An act to amend the executive law and the banking law, in relation to establishing the banking bill of rights PURPOSE: To require banks to provide consumers with their reasons for denying a debit or credit card application or account closure, and to provide an opportunity for consumers to contest or correct the information used in such adverse actions. SUMMARY OF SPECIFIC PROVISIONS: Section 1 sets forth the legislative intent.