Bill S7593 version original - Jake Ashby
BILL NUMBER: S7593REVISED 7/12/23 SPONSOR: ASHBY TITLE OF BILL: An act to amend the tax law, in relation to establishing the retire strong tax credit PURPOSE OR GENERAL IDEA OF BILL: This bill would create a property tax credit for individuals 65 years of age and older. The credit would equal 50% of real property taxes paid, not to exceed $6,500.