Bill S7596 version original - Dean Murray
BILL NUMBER: S7596 SPONSOR: MURRAY TITLE OF BILL: An act to amend the penal law, in relation to establishing the offense of obstructing a police officer PURPOSE: Makes it a felony to intentionally interfere with or inhibit a police officer or peace officer from engaging in the course of performing their official duties. SUMMARY OF PROVISIONS: Section One: Adds a new Section 195.