Bill S7598B version original - Rachel May
BILL NUMBER: S7598 SPONSOR: MAY TITLE OF BILL: An act to amend the public authorities law, in relation to the board of directors of the Central New York Regional Market Authority PURPOSE: To improve the governance and transparency to the public of the Central New York Regional Market Authority. SUMMARY OF PROVISIONS: Section 1: Amends Section 827 of the Public Authorities Law, making a technical amendment.