Bill S6110C version B - George Borrello
BILL NUMBER: S6110B SPONSOR: BORRELLO TITLE OF BILL: An act to amend the public officers law, in relation to waiving the local residency requirement for the village clerk and village treasurer in the village of Portville PURPOSE OR GENERAL IDEA OF BILL: Waives the local residency requirement for the village clerk and village treasurer in the village of Portville. SUMMARY OF PROVISIONS: Section 1.