Bill S7611 version original - James Sanders Jr.
BILL NUMBER: S7611 SPONSOR: SANDERS TITLE OF BILL: An act to amend the agriculture and markets law, in relation to the confinement of animals for food producing purposes PURPOSE: This bill phases out pig gestation crates, veal crates and hen battery cages. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the agriculture and markets law by adding a new section 353-e that defines the terms employed for the purposes of this bill.