Bill S4066A version A - John Mannion
BILL NUMBER: S4066A SPONSOR: MANNION TITLE OF BILL: An act to amend the general business law, the penal law, the general municipal law and the executive law, in relation to the registration of home improvement contractors PURPOSE OF BILL: The purpose of this bill is to reduce fraudulent acts by home improve- ment contractors by requiring contractors to register with the state. SUMMARY OF PROVISIONS: Section 1 of this bill would establish the "Home Improvement Fraud Prevention Act".