Bill S7624 version original - Steve Rhoads
BILL NUMBER: S7624 SPONSOR: RHOADS TITLE OF BILL: An act to amend the workers' compensation law, in relation to certain claims for compensation based on a COVID-19 diagnosis; and providing for the repeal of such provisions upon the expiration thereof PURPOSE: TO amend the workers compensation law, in relation to certain claims for compensation based on a COVID-19 diagnosis; and providing for the repeal of such provisions upon the expiration thereof. SUMMARY OF PROVISIONS: Section 1.