Bill S7628 version original - Rachel May
BILL NUMBER: S7628 SPONSOR: MAY TITLE OF BILL: An act to amend the executive law, in relation to establishing a Finger Lakes commission PURPOSE:: This bill establishes a Finger Lakes commission to assist local govern- ments and organizations in preserving the resources of the region for future generations. SUMMARY OF PROVISIONS:: This bill would add new sections 848-a through 848-e to the executive law.