Bill S7630 version original - Timothy M. Kennedy
BILL NUMBER: S7630 SPONSOR: KENNEDY TITLE OF BILL: An act to extend the duration of certain brownfield redevelopment and remediation tax credits for certain sites PURPOSE OR GENERAL IDEA OF BILL:: This bill would extend the duration of the brownfield redevelopment and remediation tax credits for a site in Lackawanna, Erie County to promote economic rejuvenation and decontamination. SUMMARY OF SPECIFIC PROVISIONS:: Section 1.