Bill S7643 version original - Peter Oberacker
BILL NUMBER: S7643 SPONSOR: OBERACKER TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to the opera- tion of bicycles with electric assist PURPOSE: To allow electric assist bicycles to ride on the same roads that regular bicycles are allowed to ride on. SUMMARY OF PROVISIONS: To amend Subdivision 8 of section 1242 of the vehicle and traffic law, as added by section 8 of part XX of the laws of 2020.