Bill S310A version A - Julia Salazar
BILL NUMBER: S310A SPONSOR: SALAZAR TITLE OF BILL: An act to amend the education law and the public health law, in relation to the practice of certified professional midwifery PURPOSE OR GENERAL IDEA OF BILL: To create the profession of licensed certified professional midwife. SUMMARY OF PROVISIONS: Section 1 amends Title 8 of the Education Law by adding a new Article 140-a to establish the profession of certified professional midwifery.