Bill S7661 version original - Mark Walczyk
BILL NUMBER: S7661 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the penal law, in relation to the crime of criminal impersonation in the first degree PURPOSE:: This bill would explicitly add impersonating elections officials to the crime of criminal impersonation. SUMMARY OF PROVISIONS:: Section 1: Adds subsection 2(a) and (b) to add impersonating elections officials to the crime of criminal impersonation and creates the circum- stances under which such person may be charged with the crime.