Bill S7673 version original - Mario Mattera
BILL NUMBER: S7673 SPONSOR: MATTERA TITLE OF BILL: An act in relation to protecting the health insurance benefits of reti- rees of public employers and contributions of retirees of public employ- ers PURPOSE OR GENERAL IDEA OF BILL: This bill prohibits public employers from diminishing the health insur- ance benefits of public employee retirees. SUMMARY OF PROVISIONS: Section 1 establishes that the name of the act shall be the "Health Equity for Retirees Act.