Bill S3596C version B - Pete Harckham
BILL NUMBER: S3596B SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the tax law, in relation to residential solar tax cred- its PURPOSE: The purpose of this bill is to increase the maximum credit amount for the personal income tax credit for residential solar energy systems. SUMMARY OF PROVISIONS: Section 1. subsection (g-1) of section 606 of the tax law, is amended to: -Increase the tax credit maximum to $10,000 for equipment installed on or after January first, two thousand twenty-five.