Bill S7676B version original - Jessica Ramos
BILL NUMBER: S7676 SPONSOR: RAMOS TITLE OF BILL: An act to amend the general obligations law, in relation to contracts for the creation and use of digital replicas PURPOSE: Amends general obligations law to ensure that workers have proper repre- sentation, by legal counsel of a labor organization, when negotiating the use of digital recreation and thus protecting a performer's digital voice and likeness. SUMMARY OF PROVISIONS: Section 1.