Bill S7681 version original - Toby Ann Stavisky
BILL NUMBER: S7681 SPONSOR: STAVISKY TITLE OF BILL: An act to amend the penal law, the domestic relations law and the family court act, in relation to termination of parental rights in cases of a child conceived through rape PURPOSE: To expand the protections afforded to women who conceive a child as a result of rape. SUMMARY OF LEGISLATION: SECTION 1. SECTION 130.92 OF THE PENAL LAW IS AMENDED BY ADDING A NEW SUBDIVISION 4.