Bill S7688 version original - James Tedisco
BILL NUMBER: S7688 SPONSOR: TEDISCO TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to requiring the installation and use of security camera systems at all entrances and exits of state parks PURPOSE: To require the installation and use of security camera systems at all entrances and exits of state parks. SUMMARY OF PROVISIONS: § 1. Section 3.