Bill S7700A version original - Lea Webb
BILL NUMBER: S7700 SPONSOR: WEBB TITLE OF BILL: An act to amend the social services law, in relation to authorizing a prenatal and postpartum informational mobile application PURPOSE: Authorizes a prenatal and postpartum informational mobile application SUMMARY OF PROVISIONS: Section 1 directs the commissioner to contract with a platform provider for a New York specific, membership-based mobile application to be made available to birthing individuals who are eligible for Medicaid. Section 2.