Bill S7717A version original - Robert Rolison
BILL NUMBER: S7717 SPONSOR: ROLISON TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to establishing the offense of aggravated reckless driving PURPOSE: To establish the offense of aggravated reckless driving and to allow the seizure and forfeiture of vehicles that drive recklessly. SUMMARY OF PROVISIONS: Section 1 amends the vehicle and traffic law by Adding a new section. § 1212-a Aggravated Reckless Driving. 1.